fix bug for maven.sh

+++ src/builder/maven.sh
@@ -8,7 +8,11 @@ remove_plugin() {
   else
     # 如果有参数，执行命令
     for param in "$@"; do
-      python3 /usr/share/java-utils/pom_editor.py pom_remove_plugin -r :"$param"
+      non_space_param=$(echo "$param" | tr -d '[:space:]')
+      if [[ -n "$non_space_param" ]]; then
+        echo "python3 /usr/share/java-utils/pom_editor.py pom_remove_plugin -r :$non_space_param"
+        python3 /usr/share/java-utils/pom_editor.py pom_remove_plugin -r :"$non_space_param"
+      fi
     done
   fi
 }
@@ -20,7 +24,11 @@ disable_module() {
   else
     # 如果有参数，执行命令
     for param in "$@"; do
-      python3 /usr/share/java-utils/pom_editor.py pom_disable_module -r :"$param"
+      non_space_param=$(echo "$param" | tr -d '[:space:]')
+      if [[ -n "$non_space_param" ]]; then
+        echo "python3 /usr/share/java-utils/pom_editor.py pom_disable_module -r :$non_space_param"
+        python3 /usr/share/java-utils/pom_editor.py pom_disable_module -r :"$non_space_param"
+      fi
     done
   fi
 }
@@ -32,7 +40,11 @@ delete_dir() {
   else
     # 如果有参数，执行命令
     for param in "$@"; do
-      rm -rf "$param"
+      non_space_param=$(echo "$param" | tr -d '[:space:]')
+      if [[ -n "$non_space_param" ]]; then
+        echo "rm -rf $param"
+        rm -rf "$param"
+      fi
     done
   fi
 }
